chore: ignore coverage.html output file
