+++ pkg/classifier/classifier_test.go
@@ -1959,22 +1959,3 @@ func TestClassify_WgetOutput_Escalate(t *testing.T) {
 		}
 	}
 }
-
-func TestClassify_DailyBucketAutoApproveRate(t *testing.T) {
-	b := DailyBucket{
-		Date:      "2026-04-13",
-		AutoAllow: 8,
-		AutoDeny:  1,
-		Escalate:  1,
-		Total:     10,
-	}
-	got := b.AutoApproveRate()
-	if got != 0.8 {
-		t.Errorf("AutoApproveRate() = %v, want 0.8", got)
-	}
-
-	empty := DailyBucket{}
-	if empty.AutoApproveRate() != 0 {
-		t.Errorf("AutoApproveRate() on zero bucket = %v, want 0", empty.AutoApproveRate())
-	}
-}
